chore(vmd): refactor sdk/framework/two_phase_reconciler concepts
 - ReconcilerCore — this is reconciler implementing controller-runtime related hooks and common routines.

- TwoPhaseReconciler — this is userspace reconciler, implementing Sync, UpdateStatus and some other methods for actual reconciler logic.

- ReconcilerState is an arbitrary struct implementing corresponding interface.
