+++ pkg/sdk/framework/two_phase_reconciler/reconciler_state.go
@@ -12,12 +12,13 @@ type ReconcilerState interface {
 
 	SetReconcilerResult(result *reconcile.Result)
 	GetReconcilerResult() *reconcile.Result
+
 	Reload(ctx context.Context, req reconcile.Request, log logr.Logger, client client.Client) error
+	ShouldReconcile() bool
+	ShouldApplyUpdateStatus() bool
 }
 
 type ReconcilerStateApplier interface {
 	ApplySync(ctx context.Context, log logr.Logger) error
 	ApplyUpdateStatus(ctx context.Context, log logr.Logger) error
-	ShouldApplyUpdateStatus() bool
-	ShouldReconcile() bool
 }
