+++ pkg/sdk/framework/two_phase_reconciler/two_phase_reconciler.go
@@ -1,0 +1,15 @@
+package two_phase_reconciler
+
+import (
+	"context"
+	"sigs.k8s.io/controller-runtime/pkg/controller"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+)
+
+type TwoPhaseReconciler[T ReconcilerState] interface {
+	SetupController(ctx context.Context, mgr manager.Manager, ctr controller.Controller) error
+	Sync(ctx context.Context, req reconcile.Request, state T, opts ReconcilerOptions) error
+	UpdateStatus(ctx context.Context, req reconcile.Request, state T, opts ReconcilerOptions) error
+	NewReconcilerState(opts ReconcilerOptions) T
+}
